Relocate type and implement accordingly
